Rename src/lib to src/helpers
