+++ packages/colony-example-vue/src/api/funding.js
@@ -1,6 +1,6 @@
 import { BN } from 'web3-utils'
-import { getColony } from '@/lib/colony-store'
-import sendProxy from '@/lib/send-proxy'
+import { getColony } from '@/helpers/colony-store'
+import sendProxy from '@/helpers/send-proxy'
 
 export function moveFunds({
   colony = getColony(), fromPot, toPot, amount,
