+++ packages/colony-example-vue/src/helpers/colony-store.js
