+++ packages/colony-example-vue/src/helpers/db.js
