+++ packages/colony-example-vue/src/helpers/ecp.js
