+++ packages/colony-example-vue/src/helpers/multisig.js
