+++ packages/colony-example-vue/src/helpers/network-store.js
