+++ packages/colony-example-vue/src/helpers/send-proxy.js
