+++ packages/colony-example-vue/src/views/manage/colony/tasks/Create.vue
@@ -28,7 +28,7 @@ export default {
       await this.hydrateTasks()
     },
     ...mapActions({
-      'hydrateTasks': 'tasks/hydrate',
+      hydrateTasks: 'tasks/hydrate',
     }),
   },
 }
