+++ packages/colony-starter-vue/src/helpers/ecp.js
