+++ packages/colony-starter-vue/src/helpers/multisig.js
